altered the file type of the header files from .h -> .hpp, and separated the vfx and sfx assets
